requirements.txt
fastapi
anthropic
nltk
python-dotenv